refactor: move prettyjson under credentials
